Add DNDObject sublcass
