+++ src/main.py
@@ -3,9 +3,9 @@ import logging
 import discord
 import os
 from dice import DiceExpression, DiceEmbed, RollMode
-from items import ItemEmbed, ItemList, MultiItemSelectView, NoItemsFoundEmbed
-from search import NoSearchResultsFoundEmbed, SearchEmbed, search_from_query
-from spells import MultiSpellSelectView, NoSpellsFoundEmbed, SpellEmbed, SpellList
+from dnd import ItemList, SpellList
+from embeds import ItemEmbed, MultiItemSelectView, MultiSpellSelectView, NoItemsFoundEmbed, NoSearchResultsFoundEmbed, NoSpellsFoundEmbed, SpellEmbed
+from search import SearchEmbed, search_from_query
 from discord import app_commands
 from dotenv import load_dotenv
 from stats import StatsEmbed
